Upload of Code
At this point the grid can transform a [500][500] grid of integers into a display to be seen. ( This is so we can map the zone we are traveling in ) . Also have some shell classes and some of the basic impelementation structure that will be included in the Main.java file.